pyproject.toml
[project]
name = "pdf-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "inngest>=0.4.22",
    "llama-index-core>=0.14.8",
    "llama-index-readers-file>=0.5.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "qdrant-client==1.15.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
